requirements.txt
channels
uvicorn
backports.tarfile
celery
django-ckeditor-5
django-cors-headers
django-environ
django-extensions
django-filter
django-modeltranslation
django-redis
django-rosetta
django-unfold==0.42.0
djangorestframework-simplejwt
drf-spectacular
importlib-metadata
importlib-resources
jaraco.collections
packaging
pip-chill
platformdirs
psycopg2-binary
tomli
django-silk==5.3.1
django-storages
boto3
